backend/requirements.txt
databases==0.9.0
fastapi==0.115.7
bcrypt==4.2.1
SQLAlchemy==2.0.37
starlette==0.45.3
uvicorn==0.34.0
watchdog==6.0.0
pymysql==1.1.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
